--- src/main/java/com/example/lawComponents/LawParsingApplication.java
@@ -1,11 +1,0 @@
-package com.example.lawComponents;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LawParsingApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(LawParsingApplication.class, args);
-    }
-}
