+++ src/main/java/com/example/pracrawling/entity/Addendum.java
@@ -1,4 +1,4 @@
-package com.example.lawComponents.addendum;
+package com.example.pracrawling.entity;
 
 import com.example.pracrawling.entity.Law;
 import lombok.AllArgsConstructor;
