+++ src/main/java/com/example/pracrawling/repository/AddendumRepository.java
@@ -1,5 +1,6 @@
-package com.example.lawComponents.addendum;
+package com.example.pracrawling.repository;
 
+import com.example.pracrawling.entity.Addendum;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AddendumRepository extends JpaRepository<Addendum, Long> {
