+++ src/main/java/com/example/pracrawling/repository/AmendmentRepository.java
@@ -1,5 +1,6 @@
-package com.example.lawComponents.amendment;
+package com.example.pracrawling.repository;
 
+import com.example.pracrawling.entity.Amendment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AmendmentRepository extends JpaRepository<Amendment, Long> {
